RS Post Settings: Extract URI-to-file logic into UriToFileMapper
Move copyUriToTempFile and its Context dependency out of the
ViewModel into a dedicated injectable helper. This improves
testability and removes the @ApplicationContext from the ViewModel.
